+++ migrations/06_rss_optional.php
@@ -1,15 +1,15 @@
 <?php
 class RssOptional extends Migration
 {
-    public function description ()
+    public function description()
     {
         return 'config eintraege fuer rss werden angelegt';
     }
 
-    public function up ()
+    public function up()
     {
         Config::get()->create('BULLETIN_BOARD_ENABLE_RSS', [
-            'value'       => true,
+            'value'       => (int) true,
             'type'        => 'boolean',
             'range'       => 'global',
             'section'     => 'SchwarzesBrettPlugin',
@@ -17,7 +17,7 @@ class RssOptional extends Migration
         ]);
     }
 
-    public function down ()
+    public function down()
     {
         Config::get()->delete('BULLETIN_BOARD_ENABLE_RSS');
     }
